src/steplist_ai_features/signatures/emojifier.py
"""Signature definitions for emojifier."""

import dspy

class EmojifierSignature(dspy.Signature):
    """
    """

    title: str = dspy.InputField(desc="")
    emoji: str = dspy.OutputField(desc="")
